followされた際の通知機能を切り出すためのmoduleの導入
