+++ app/models/user.rb
@@ -1,6 +1,8 @@
 class User < ApplicationRecord
   # Include default devise modules. Others available are:
   # :confirmable, :lockable, :timeoutable, :trackable and :omniauthable
+  include CreateFollowNotificationService
+  
   devise :database_authenticatable, :registerable,
          :recoverable, :rememberable, :validatable
   has_one_attached :image
@@ -75,14 +77,4 @@ class User < ApplicationRecord
     end
   end
 
-  def create_notification_follow!(current_user)
-    temp = Notification.where(['visitor_id = ? and visited_id = ? and action = ? ', current_user.id, id, 'follow'])
-    return unless temp.blank?
-
-    notification = current_user.active_notifications.new(
-      visited_id: id,
-      action: 'follow'
-    )
-    notification.save if notification.valid?
-  end
 end
