+++ app/services/create_follow_notification_service.rb
@@ -1,0 +1,12 @@
+module CreateFollowNotificationService
+  def create_notification_follow!(current_user)
+    temp = Notification.where(['visitor_id = ? and visited_id = ? and action = ? ', current_user.id, id, 'follow'])
+    return unless temp.blank?
+
+    notification = current_user.active_notifications.new(
+      visited_id: id,
+      action: 'follow'
+    )
+    notification.save if notification.valid?
+  end
+end
